+++ statistics/scalar.go
@@ -102,19 +102,19 @@ func (hg *Histogram) PreCalculateScalar() {
 	}
 	switch hg.GetLower(0).Kind() {
 	case types.KindMysqlDecimal, types.KindMysqlTime:
-		hg.scalars = make([]scalar, l)
+		hg.Scalars = make([]scalar, l)
 		for i := 0; i < l; i++ {
-			hg.scalars[i] = scalar{
+			hg.Scalars[i] = scalar{
 				lower: convertDatumToScalar(hg.GetLower(i), 0),
 				upper: convertDatumToScalar(hg.GetUpper(i), 0),
 			}
 		}
 	case types.KindBytes, types.KindString:
-		hg.scalars = make([]scalar, l)
+		hg.Scalars = make([]scalar, l)
 		for i := 0; i < l; i++ {
 			lower, upper := hg.GetLower(i), hg.GetUpper(i)
 			common := commonPrefixLength(lower.GetBytes(), upper.GetBytes())
-			hg.scalars[i] = scalar{
+			hg.Scalars[i] = scalar{
 				commonPfxLen: common,
 				lower:        convertDatumToScalar(lower, common),
 				upper:        convertDatumToScalar(upper, common),
@@ -137,9 +137,9 @@ func (hg *Histogram) calcFraction(index int, value *types.Datum) float64 {
 	case types.KindMysqlDuration:
 		return calcFraction(float64(lower.GetDuration(0, 0).Duration), float64(upper.GetDuration(0, 0).Duration), float64(value.GetMysqlDuration().Duration))
 	case types.KindMysqlDecimal, types.KindMysqlTime:
-		return calcFraction(hg.scalars[index].lower, hg.scalars[index].upper, convertDatumToScalar(value, 0))
+		return calcFraction(hg.Scalars[index].lower, hg.Scalars[index].upper, convertDatumToScalar(value, 0))
 	case types.KindBytes, types.KindString:
-		return calcFraction(hg.scalars[index].lower, hg.scalars[index].upper, convertDatumToScalar(value, hg.scalars[index].commonPfxLen))
+		return calcFraction(hg.Scalars[index].lower, hg.Scalars[index].upper, convertDatumToScalar(value, hg.Scalars[index].commonPfxLen))
 	}
 	return 0.5
 }
