--- statistics/selectivity.go
@@ -1,44 +1,0 @@
-// Copyright 2017 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package statistics
-
-import (
-	"github.com/pingcap/tidb/util/ranger"
-)
-
-// StatsNode is used for calculating selectivity.
-// TODO: remove this structure after removing tidb_optimizer_selectivity_level.
-type StatsNode struct {
-	// Ranges contains all the Ranges we got.
-	Ranges []*ranger.Range
-	Tp     int
-	ID     int64
-	// mask is a bit pattern whose ith bit will indicate whether the ith expression is covered by this index/column.
-	mask int64
-	// Selectivity indicates the Selectivity of this column/index.
-	Selectivity float64
-	// numCols is the number of columns contained in the index or column(which is always 1).
-	numCols int
-	// partCover indicates whether the bit in the mask is for a full cover or partial cover. It is only true
-	// when the condition is a DNF expression on index, and the expression is not totally extracted as access condition.
-	partCover bool
-}
-
-// The type of the StatsNode.
-const (
-	IndexType = iota
-	PkType
-	ColType
-)
