--- statistics/testdata/trace_suite_in.json
@@ -1,19 +1,0 @@
-[
-  {
-    "name": "TestTraceCE",
-    "cases": [
-      "a > 0 and a < 2",
-      "a >= 1 and a < 10",
-      "a < 3 or b < 4",
-      "a = 1 and b = 2",
-      "a < 5 or a > 10"
-    ]
-  },
-  {
-    "name": "TestTraceDebugSelectivity",
-    "cases": [
-      "select * from t where a = 100 and b = 350",
-      "select * from t where a < -1500 and b > 400 and b < 403"
-    ]
-  }
-]
